requirements/dev.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements/dev.in
#
asgiref==3.5.0
    # via django
certifi==2021.10.8
    # via requests
cfgv==3.3.1
    # via pre-commit
charset-normalizer==2.0.11
    # via requests
click==8.0.3
    # via pip-tools
deprecated==1.2.13
    # via redis
distlib==0.3.4
    # via virtualenv
django==4.0.2
    # via
    #   -r requirements/base.in
    #   django-safedelete
django-environ==0.8.1
    # via -r requirements/base.in
django-safedelete==1.1.2
    # via -r requirements/base.in
filelock==3.4.2
    # via virtualenv
identify==2.4.7
    # via pre-commit
idna==3.3
    # via requests
kavenegar==1.1.2
    # via -r requirements/base.in
nodeenv==1.6.0
    # via pre-commit
packaging==21.3
    # via redis
pep517==0.12.0
    # via pip-tools
pillow==9.0.0
    # via -r requirements/base.in
pip-tools==6.1.0
    # via -r requirements/dev.in
platformdirs==2.4.1
    # via virtualenv
pre-commit==2.16.0
    # via -r requirements/dev.in
psycopg2==2.9.3
    # via -r requirements/base.in
pyparsing==3.0.7
    # via packaging
pyyaml==6.0
    # via pre-commit
redis==4.1.2
    # via -r requirements/base.in
requests==2.27.1
    # via kavenegar
six==1.16.0
    # via virtualenv
sqlparse==0.4.2
    # via django
toml==0.10.2
    # via pre-commit
tomli==2.0.0
    # via pep517
urllib3==1.26.8
    # via requests
virtualenv==20.13.0
    # via pre-commit
wrapt==1.13.3
    # via deprecated

# The following packages are considered to be unsafe in a requirements file:
# pip
